src: Move MockedSourceGroup to new file
